change sa_flags in sigaction to c_int on redox
(backport <https://github.com/rust-lang/libc/pull/4986>)
(cherry picked from commit 939f0090c2190755b0617172ba1b8ddbc5077d20)

+++ src/unix/redox/mod.rs
@@ -153,7 +153,7 @@ s! {
     #[allow(unpredictable_function_pointer_comparisons)]
     pub struct sigaction {
         pub sa_sigaction: crate::sighandler_t,
-        pub sa_flags: c_ulong,
+        pub sa_flags: c_int,
         pub sa_restorer: Option<extern "C" fn()>,
         pub sa_mask: crate::sigset_t,
     }
